pyproject.toml
[tool.poetry]
name = "trivia"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Flask = "^2.3.2"
flask-restful = "^0.3.10"
python-dotenv = "^1.0.0"
flask-sqlalchemy = "^3.0.5"
flask-migrate = "^4.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
